LPS-161984 Extract SearchableContainer from EditObjectField to its own component
